more stuff that will certainly work nocap
