+++ scripts/caravan/caravan-outpost-gui.lua
@@ -1,24 +1,7 @@
 require 'caravan-gui-shared'
 local Table = require('__stdlib__/stdlib/utils/table')
 
-local function has_schedule(caravan_data, entity)
-    if not Caravan.validity_check(caravan_data) then
-        return
-    end
-    if not caravan_data.schedule then
-        return
-    end
-    return Table.any(caravan_data.schedule, function(schedule)
-        return schedule.entity == entity
-    end)
-end
 
--- probably bad if you have 10 k caravans... but is that even remotely relevant?
-local function has_any_caravan(entity)
-    return Table.any(global.caravans, function(caravan_data)
-        return has_schedule(caravan_data, entity)
-    end)
-end
 
 Caravan.events.on_gui_opened_outpost = function(event)
     local player = game.get_player(event.player_index)
